requirements.txt
# Core Framework
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Database
sqlalchemy>=2.0.25,<3.0.0
asyncpg>=0.29.0,<1.0.0
alembic>=1.13.0,<2.0.0

# LLM Integration
anthropic>=0.40.0,<1.0.0

# Template Engine
jinja2>=3.1.3,<4.0.0

# Task Queue
celery>=5.3.6,<6.0.0
redis>=5.0.1,<6.0.0

# Email & Distribution
sendgrid>=6.11.0,<7.0.0

# HTTP Client
httpx>=0.26.0,<1.0.0

# Content Processing
markdown>=3.5.2,<4.0.0
beautifulsoup4>=4.12.3,<5.0.0
weasyprint>=61.0,<62.0

# Configuration
pyyaml>=6.0.1,<7.0.0

# Text Analysis
textstat>=0.7.3,<1.0.0

# Testing
pytest>=7.4.4,<9.0.0
pytest-asyncio>=0.23.3,<1.0.0
pytest-cov>=4.1.0,<6.0.0
httpx>=0.26.0,<1.0.0

# Development
ruff>=0.2.0,<1.0.0
mypy>=1.8.0,<2.0.0
